+++ generators/service-mongodb/templates/swift/instrumentation.swift
@@ -6,7 +6,7 @@ func initializeServiceMongodb(cloudEnv: CloudEnv) throws -> Database {
     guard let mongodbCredentials = cloudEnv.getMongoDBCredentials(name: "{{servLookupKey}}") else {
         throw InitializationError("Could not load credentials for MongoDB.")
     }
-    let mongodb = try Database(mongodbCredentials.uri)
+    let mongodb = try Database.synchronousConnect(mongodbCredentials.uri)
     Log.info("Found and loaded credentials for MongoDB.")
     return mongodb
 }